Update patch set 1
Patch Set 1:

This will show the complete configuration of all checkers in the job logs (not only those related to Zuul).  I did that intentionally to aid contributors who may not otherwise have access to the list checkers endpoint (since it requires the "Administrate checkers" permission).  Outputting it here makes it easily accessible so that people can identify existing checkers and alter their configuration.  I don't believe there is any sensitive information there, but it would be good for a Gerrit Maintainer to confirm that is the case before we merge this.

Patch-set: 1