Limit Nodepool to 4 servers
It looks like when we try to create more than 4 servers at once
it fails.  I'm unsure why (is there a quota? on what resource?).

If there is a quota, we should be able to have nodepool handle it
automatically.  In the mean time, if we tell nodepool not to create
so many servers, we should be able to avoid this problem.

Change-Id: Ic7d08593889891ded4aec8114594c068f5330f42