Update patch set 2
Patch Set 2:

Not yet; we should probably add a dashboard to collect usage metrics.  I did a quick test involving a complete build and test of the gerrit master branch and core plugins, and that took more than 2 hours with 2cpus, 46 minutes with 8 cpus, and 33 minutes with 16 cpus.  So we probably want 16cpus for that kind of job.  We can probably use smaller machines for the plugin build jobs (and once we add GCE quota support to nodepool, we'll be able to lift the 4-server restriction and nodepool will handle the mixed sizes automatically).

In short, I would say let's not request a larger quota right now -- let's see if we end up with a backlog with the current allocation first and collect more data to make sure we have a good justification for asking for a larger quota.

Patch-set: 2