Disable host key checking in nodepool
We're seeing a lot of builds fail with RETRY errors.  This is due
to the host key on the remote node changing between the time that
nodepool collects it (immediately at boot), and Zuul uses it.  It
appears that the images we're using have an ssh host key baked in
to them but then overwrite it at first boot.  Because we're using
the internal IP (which doesn't need to wait for NAT setup), Nodepool
can end up connecting to the VM very quickly and grabbing the
initial (as opposed to the newly generated host key).  When Zuul
tries to talk to the node, it has changed and it throws an error.

By disabling this, we allow Zuul to simply accept the host key on
first connection.  This is low-risk since we're only using the
internal cloud network anyway.

Change-Id: I2b2347f37527a72866e92164e6779ab6b57ebfa7