Use correct zk environment variable names
There seem to be two standards for env variable names and we were
using the wrong ones for the actual startup script that is used.
Switch the snapshot retention configuration to use the correct names
and remove some others where we are currently using the default
values.

This should correct the issue with the snapshots filling the disk.

Change-Id: Ib5176838e40265d24500f50775911abca7ae0fb5